requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.0
django-environ==0.11.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
Pillow==10.1.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
